initfn abi fix
