+++ src/detail/mod.rs
@@ -36,8 +36,9 @@ pub mod asm;
 #[cfg(all(unix, target_arch = "loongarch64"))]
 #[path = "loongarch64_unix.rs"]
 pub mod asm;
+mod gen;
 
-pub use self::asm::{initialize_call_frame, prefetch, swap_registers, Registers};
+pub use self::asm::{gen_init, initialize_call_frame, prefetch, swap_registers, InitFn, Registers};
 
 #[inline]
 fn align_down(sp: *mut usize) -> *mut usize {
